Rename the Component from assignment to assignments
