+++ src/router/index.ts
@@ -3,8 +3,8 @@ import VueRouter, { RouteConfig } from "vue-router";
 import Index from "@/views/Index.vue";
 import Home from "@/views/Home.vue";
 import SignIn from "@/views/SignIn.vue";
-import AssignmentView from "@/views/Assignment.vue";
-import Assignments from "@/views/Assignments.vue"
+import AssignmentDetailView from "@/views/Assignment/AssignmentDetailView.vue";
+import Assignments from "@/views/Assignment/Assignments.vue";
 import Account from "@/views/Account.vue";
 
 Vue.use(VueRouter);
@@ -28,12 +28,12 @@ const routes: Array<RouteConfig> = [
   {
     path: "/assignment",
     name: "Assignments",
-    component: Assignments
+    component: Assignments,
   },
   {
     path: "/assignment/:id",
-    name: "AssignmentPage",
-    component: AssignmentView,
+    name: "AssignmentDetailPage",
+    component: AssignmentDetailView,
   },
   {
     path: "/account",
