+++ src/views/Assignment/AssignmentDetailView.vue
